requirements.txt
svgwrite
tqdm
request